requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==3.4.8
defusedxml==0.7.1
Django==3.2.7
django-allauth==0.45.0
django-better-admin-arrayfield==1.4.2
django-cors-headers==3.8.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.1.0
idna==3.2
oauthlib==3.1.1
Pillow==8.3.2
psycopg-binary==3.0.3
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
